+++ examples/blog/theme.config.jsx
@@ -1,5 +1,18 @@
 /* eslint sort-keys: error */
 export default {
+  components: {
+    h1: ({ children }) => (
+      <h1 style={{
+        WebkitBackgroundClip: 'text',
+        WebkitTextFillColor: 'transparent',
+        backgroundClip: 'text',
+        backgroundImage: 'linear-gradient(90deg,#7928CA,#FF0080)',
+      }}
+      >
+        {children}
+      </h1>
+    )
+  },
   darkMode: true,
   footer: (
     <small style={{ display: 'block', marginTop: '8rem' }}>
@@ -9,12 +22,13 @@ export default {
       >
         CC BY-NC 4.0
       </abbr>{' '}
-      <time>{new Date().getFullYear()}</time> © Shu Ding.
+      {new Date().getFullYear()} © Shu Ding.
       <a href="/feed.xml">RSS</a>
       <style jsx>{`
         a {
           float: right;
         }
+
         @media screen and (max-width: 480px) {
           article {
             padding-top: 2rem;
@@ -23,5 +37,5 @@ export default {
         }
       `}</style>
     </small>
-  ),
+  )
 }